+++ test/index.mjs
@@ -1,5 +1,4 @@
 import TestDirector from "test-director";
-import test_bundle from "./bundle.test.mjs";
 import test_createUploadLink from "./public/createUploadLink.test.mjs";
 import test_formDataAppendFile from "./public/formDataAppendFile.test.mjs";
 
@@ -7,6 +6,5 @@ const tests = new TestDirector();
 
 test_createUploadLink(tests);
 test_formDataAppendFile(tests);
-test_bundle(tests);
 
 tests.run();