+++ test/public/formDataAppendFile.test.mjs
@@ -1,9 +1,17 @@
 import { strictEqual } from "assert";
 import { File, FormData } from "formdata-node";
 import formDataAppendFile from "../../public/formDataAppendFile.js";
+import assertBundleSize from "../assertBundleSize.mjs";
 
 export default (tests) => {
-  tests.add("`formDataAppendFile`.", () => {
+  tests.add("`formDataAppendFile` bundle size.", async () => {
+    await assertBundleSize(
+      new URL("../../public/formDataAppendFile.js", import.meta.url),
+      150
+    );
+  });
+
+  tests.add("`formDataAppendFile` functionality.", () => {
     const formData = new FormData();
     const fieldName = "a";
     const fileName = "a.txt";